Removed option to search fuel transactions
Unless they're searching for a specific date, there's no reason to search. You can sort in the ManageFuelupsPage. In 3 years, I only made 126 fuelups, at slightly less than the average of 12,000 miles a year. If you're searching for a specific place or a date, you can sort it and find it fairly quickly, I think.